consolidate & dedupe send selectors
