+++ ui/app/components/app/modals/account-details-modal/account-details-modal.container.js
@@ -1,6 +1,9 @@
 import { connect } from 'react-redux'
 import { showModal, setAccountLabel } from '../../../../store/actions'
-import { getSelectedIdentity, getRpcPrefsForCurrentProvider } from '../../../../selectors'
+import {
+  getSelectedIdentity,
+  getRpcPrefsForCurrentProvider,
+} from '../../../../selectors'
 import AccountDetailsModal from './account-details-modal.component'
 
 const mapStateToProps = (state) => {
