+++ ui/app/components/ui/currency-input/currency-input.container.js
@@ -1,8 +1,12 @@
 import { connect } from 'react-redux'
 import CurrencyInput from './currency-input.component'
 import { ETH } from '../../../helpers/constants/common'
-import { getSendMaxModeState } from '../../../selectors/send'
-import { getIsMainnet, preferencesSelector } from '../../../selectors'
+import {
+  getSendMaxModeState,
+  getIsMainnet,
+  preferencesSelector,
+} from '../../../selectors'
+
 
 const mapStateToProps = (state) => {
   const { metamask: { nativeCurrency, currentCurrency, conversionRate } } = state
