+++ ui/app/ducks/alerts/unconnected-account.js
@@ -3,7 +3,10 @@ import { captureException } from '@sentry/browser'
 
 import actionConstants from '../../store/actionConstants'
 import { addPermittedAccount } from '../../store/actions'
-import { getOriginOfCurrentTab, getSelectedAddress } from '../../selectors'
+import {
+  getOriginOfCurrentTab,
+  getSelectedAddress,
+} from '../../selectors'
 
 // Constants
