+++ ui/app/helpers/higher-order-components/metametrics/metametrics.provider.js
@@ -9,7 +9,6 @@ import {
   getAccountType,
   getNumberOfAccounts,
   getNumberOfTokens,
-
   txDataSelector,
 } from '../../../selectors'
