+++ ui/app/pages/permissions-connect/permissions-connect.container.js
@@ -7,7 +7,6 @@ import {
   getAccountsWithLabels,
   getLastConnectedInfo,
   getTargetDomainMetadata,
-
   getPermissionDomains,
 } from '../../selectors'
