+++ ui/app/pages/send/account-list-item/tests/account-list-item-container.test.js
@@ -10,12 +10,10 @@ proxyquire('../account-list-item.container.js', {
       return () => ({})
     },
   },
-  '../../../selectors/send': {
+  '../../../selectors': {
     getConversionRate: () => `mockConversionRate`,
     getCurrentCurrency: () => `mockCurrentCurrency`,
     getNativeCurrency: () => `mockNativeCurrency`,
-  },
-  '../../../selectors/': {
     isBalanceCached: () => `mockBalanceIsCached`,
     preferencesSelector: ({ showFiatInTestnets }) => ({
       showFiatInTestnets,
