+++ ui/app/pages/send/send-content/add-recipient/ens-input.container.js
@@ -3,10 +3,9 @@ import {
   getCurrentNetwork,
   getSendTo,
   getSendToNickname,
-} from '../../../../selectors/send'
-import {
   getAddressBookEntry,
 } from '../../../../selectors'
+
 import { connect } from 'react-redux'
