+++ ui/app/pages/send/send-content/send-amount-row/amount-max-button/amount-max-button.container.js
@@ -5,8 +5,8 @@ import {
   getSendFromBalance,
   getTokenBalance,
   getSendMaxModeState,
-} from '../../../../../selectors/send'
-import { getBasicGasEstimateLoadingStatus } from '../../../../../selectors'
+  getBasicGasEstimateLoadingStatus,
+} from '../../../../../selectors'
 import { calcMaxAmount } from './amount-max-button.utils.js'
 import {
   updateSendAmount,
