+++ ui/app/pages/send/send-content/send-gas-row/send-gas-row.container.js
@@ -11,8 +11,6 @@ import {
   getGasLoadingError,
   gasFeeIsInError,
   getGasButtonGroupShown,
-} from '../../../../selectors/send'
-import {
   getAdvancedInlineGasShown,
   getCurrentEthBalance,
   getSelectedToken,
@@ -20,6 +18,7 @@ import {
   getRenderableEstimateDataForSmallButtonsFromGWEI,
   getDefaultActiveButtonIndex,
 } from '../../../../selectors'
+
 import {
   isBalanceSufficient,
   calcGasTotal,
