+++ ui/app/pages/send/send-content/send-gas-row/tests/send-gas-row-container.test.js
@@ -30,7 +30,7 @@ proxyquire('../send-gas-row.container.js', {
       return () => ({})
     },
   },
-  '../../../../selectors/send': {
+  '../../../../selectors': {
     getSendMaxModeState: (s) => `mockMaxModeOn:${s}`,
   },
   '../../send.utils.js': {
