+++ ui/app/pages/settings/contact-list-tab/add-contact/add-contact.container.js
@@ -5,7 +5,7 @@ import { withRouter } from 'react-router-dom'
 import { addToAddressBook, showQrScanner, qrCodeDetected } from '../../../../store/actions'
 import {
   getQrCodeData,
-} from '../../../../selectors/send'
+} from '../../../../selectors'
 
 const mapStateToProps = (state) => {
   return {
