+++ ui/app/selectors/confirm-transaction.js
@@ -12,6 +12,7 @@ import {
 import {
   sumHexes,
 } from '../helpers/utils/transactions.util'
+import { getNativeCurrency } from '.'
 
 const unapprovedTxsSelector = (state) => state.metamask.unapprovedTxs
 const unapprovedMsgsSelector = (state) => state.metamask.unapprovedMsgs
@@ -123,7 +124,6 @@ export const unconfirmedTransactionsCountSelector = createSelector(
 
 export const currentCurrencySelector = (state) => state.metamask.currentCurrency
 export const conversionRateSelector = (state) => state.metamask.conversionRate
-export const getNativeCurrency = (state) => state.metamask.nativeCurrency
 
 export const txDataSelector = (state) => state.confirmTransaction.txData
 const tokenDataSelector = (state) => state.confirmTransaction.tokenData
