+++ ui/app/selectors/index.js
@@ -3,5 +3,6 @@ export * from './custom-gas'
 export * from './first-time-flow'
 export * from './permissions'
 export * from './selectors'
+export * from './send'
 export * from './tokens'
 export * from './transactions'
